Fixed some bugs

+++ src/vector2.h
@@ -2,10 +2,11 @@
 #define VECTOR2_H__
 
 #include <stddef.h>
+#include <inttypes.h>
 
 typedef struct {
-    size_t x;
-    size_t y;
+    int64_t x;
+    int64_t y;
 } vector2_t;
 
 #endif // VECTOR2_H__
